Makefile: Use target aarch64-unknown-none-softfloat
Until now we were using aarch64-unknown-none. That means that the
compiler is allowed to use floating point instructions. That led to hard
to debug crashes because we haven't enabled floating point yet in our
kernel.

Switch to a softfloat target to prevent the compiler from generating FP
instructions behind our back.