exceptions: Implement saving and restoring of context
Define an ExceptionFrame struct with the context we need to maintain.
Implement save_context and restore_context routines in assembly that are
used to create a new instance of the struct and save the registers
before calling a Rust handler and then restore the register state after
returning from the handler and before jumping to the code that was
running before the exception was received.