Add a basic driver for the interrupt controller
Add a driver for the BMC2837 interrupt controller. The driver currently
only enables or disables a given IRQ. Also add irq.rs defining the
interrupt sources and an API for enabling/disabling IRQs so that other
code doesn't have to talk to the driver directly.

Use the new code by enabling the AUX interrupt. We won't receive
interrupts from the AUX device / mini UART yet, because we need to
configure the device to generate interrupts yet.