irq: Actually check for pending interrupts
Until now in el1_sp1_irq_handler() we assumed we ended up there due to
an AUX / mini UART due to it being the only one that we enable. Change
that by introducing a process_irqs() function that actually asks the
interrupt controller which IRQs are pending and then calls the right IRQ
handler or panics if an unexpected IRQ was raised.

Use static IRQ dispatch for now but in the future we should allow
drivers to register and unregister interrupt handlers dynamically.