uart_mini: Split irq handler in top and bottom half
We should strive to do as little as possible when we are in interrupt
context with interrupts disabled. In the RX IRQ handler of the mini UART
we currently both get and put each character. Split the irq handler in
top and bottom halves. In the top half just read all pending characters
from the RX FIFO of the device and store them to a circular buffer in
memory. In the bottom half copy the buffer and process it with
interrupts enabled.

Introduce a PENDING_ACTIONS bitfield that interrupt top halves will use
for recording pending actions. Check for any pending actions in the main
event loop and process them.