irq: Use compiler barriers in {enable,disable}_interrupts()
According to the ARM ARM section D1.4.1.1 'Accessing PSTATE fields':

"""
Writes to the PSTATE fields have side-effects on various aspects of the
PE operation. For side-effects caused by writes to a PSTATE field, all
of the following are true:
* The side-effect is guaranteed not to be visible to earlier
instructions in the Execution stream.
* The side-effect is guaranteed to be visible to later instructions in
the Execution stream.
"""

However, compiler barriers are still needed to make sure the compiler
doesn't re-order anything. Add them in enable_interrupts() and
disable_interrupts().