Turn paging on and configure an identity map
Rpi3 has a physical address space of 1GiB. Identity map this 1GiB. We
use 64KiB page granularity, and since we only need 30 bits to describe
the physical address space, 2 level of pages tables are enough (L2 and
L3). For now we pretend that the first 512 MiB are DRAM, and the
following 512 MiB are device memory (even though in reality only the top
16MiB are device memory). This lets us get a way with a single L2 page
table with 2 block descriptor entries, each mapping 512 MiB and no L3
tables.

This is temporary, we'll switch to using TTBR1 for the kernel mappings
soon.