boot.s: Make code position independent by using PC-relative addresses
In order to run the kernel with a different link address than its load
address we need to make the in boot.s position independent.

The LDR, = syntax is a pseudo-instruction that loads a pre-generated
absolute address. That absolute address won't work if we change the link
address of the kernel. Use the ADR instruction instead, which generates
a PC-relative address by taking the address of the instruction itself
and adding an offset that is encoded in the instruction to it.

Note that ADR works when the labels are within 1MiB of the PC, but that
is not a problem at the moment.

Signed-off-by: Ilias Stamatis <ilstam@mailbox.org>