Link and run the kernel from high addresses
Typically in the ARM64 architecture a kernel runs in high addresses in
the virtual address space using TTBR1_EL1 for its page tables.

This patch maps virtual addresses [0xffffffffc0000000,
0xffffffffffffffff] to physicall addresses [0, 0x3fffffff] by using the
same root page table for TTBR1_EL1 as we do for TTBR0_EL1. Since the
kernel is loaded and starts running at a load address we need low
address mappings when turning paging on initially.

Add a pre_main() function that does the initial setup and turns paging
on before jumping to main() using a high virtual address. Then disable
TTBR0_EL1 in main() so that low virtual addresses are not mapped in the
kernel any more.

Change the link address in the linker script to a high address to, from
0x80000 to 0xffffffffc0080000.