Introduce AddressVirtual and a new address crate
Now that we run the kernel with paging on we are using virtual address
rather than physical addresses to access MMIO registers. So the
MMIORegisters type as well as PERIPHERALS_BASE need to use virtual
addresses.

Introduce a new AddressVirtual type and move everything to a new address
crate. Use u64 instead of usize since we don't plan to support any
non-64 bit platform. Add a new as_virtual() function to the
AddressPhysical type so that we can easily convert kernel VAs to PAs
without walking the page tables. Since we're using an identity map at
the moment, as_virtual() doesn't need to apply an offset.