paging: Avoid using atomics before turning on the MMU
On ARM64 you can't reliably use atomics before enabling the MMU.

Case in point, while the code before this patch works on QEMU it doesn't
work on a real Raspbbery Pi 3. That is because we try to aquire a
spinlock which uses atomic operations before enabling the MMU.

Switch to using a static mut. In this instance it should be safe since
there's a single thread of execution so no data races.