Minimise LED blinking time when running on QEMU
The LED blinking introduces a delay when booting, which makes
development slower. I'm primarily developing on QEMU where the effects
of LED blinking aren't visible anyway. To make booting faster when
running on QEMU, reduce the wait time between toggling the LED on and
off from 500ms to 1ms.

Since we're here, get rid of the messages informing us about the LED
blinking action, and do the blinking first thing before we print
anything else to the console.