address: Change AddressRange type to RangePhysical
When creating a range, it'd be nice to check that the range is valid at
the time of creation by using the add() method. Also, in the near future
we'll need to call other common functions that both AddressPhysical and
AddressVirtual implement.

Normally we'd define define a trait that both AddressPhysical and
AddressVirtual implement and have the generic type T of AddressRange
implement that trait. Unfortunately, Rust doesn't support const
functions in trait implementation yet so that doesn't work.

Do the less nice thing instead which is to define a RangePhysical type.
In the future we might have to define a RangeVirtual type too and end up
with duplicated code unfortunately.

On the other hand, one benefit with the new implementation is that
RangePhysical::new() will panic if the range is not valid.