main: Define allocator_init() and add free mem regions to the allocator
Define allocator_init() which initializes the allocator with the
available memory regions. allocator_init() takes the total available RAM
range as returned by the firmware, punches a hole in it where the kernel
binary has been loaded (plus 2 pages for the stack and stack guard) and
gives the resulting 2 memory regions to the allocator.